requirements.txt
python-dotenv
groq
voicevox-client
requests